+++ R/slurp_helpers.R
@@ -4,6 +4,9 @@
 ## (or rather, I made the big function by adding together the smaller ones)
 ## for the sake of easier debugging
 
+# `total` means whether to ask the API for a total figure for each query
+# I don't need this and I suggest users keep this as "false"
+# (NB this needs to be the string "false" or "true" not logical FALSE or TRUE)
 build_recodes <- function(id_list, chunk, total = "false") {
 
   # sub-function
@@ -25,21 +28,26 @@ build_recodes <- function(id_list, chunk, total = "false") {
   }
 
 
-  # main function
+  # main part of build_recodes
+  # take a list of codes and a chunk of area codes and map along them with `create_recodes`
   map2_chr(id_list, chunk,
            ~ create_recodes(field_id = .x, area_id = .y, total = total)) %>%
     str_c(collapse = ",") %>%
     str_c("\"recodes\" : {", ., "}")
 }
 
+# build another part of the JSON query
 build_dimensions <- function(x) {
   str_c("[ \"", x, "\" ]") %>%
     str_c(., collapse = ",") %>%
     str_c("\"dimensions\" : [", ., "]")
 }
 
+# this is where it all gets brought together
 
 build_query <- function(build_list, geo_codes_chunk) {
+
+  # `build_list` is what is returned by `get_dwp_codes`
 
   # make all the parts (helper functions below, outside this function)
 
@@ -82,13 +90,18 @@ convert_geo_ids <- function(id, x) {
 }
 
 
-# extract and tidy the data from the API return
+# extract and tidy the data from each SX API return
+# dates is a list of all the data point dates returned (just the most recent 1 by default)
 pull_sx_data <- function(df, dates) {
   df %>%
     pluck("fields", "items", 1) %>%
     select(-1) %>%
+    # extract area codes
     mutate_at("uris", ~ str_replace(., "(.*:)([:alnum:]*$)", "\\2") %>% unlist()) %>%
+    # area names
     mutate_at("labels", ~ unlist(.)) %>%
+    # repeat as many times as needed (repeat each area for each item of dates)
+    # shouldn't this be map_dfr? - need to investigate what happens
     map_dfc( ~ rep(., times = length(dates))) %>%
     bind_cols(values = c(pluck(df, "cubes", 1, "values")))
 }